requirements.txt
pandas==1.2.4
xgboost==1.3.3
torch==1.8.1
scikit-learn===0.23.2
onnxruntime
onnx
seaborn
matplotlib
click==8.0.1
shap==0.39.0
lime==0.2.0.1
scikit-optimize==0.8.1
tqdm==4.60.0
onnxmltools==1.7.0
hummingbird-ml==0.3.1